Neural network ready for training
Fixed win checking script. It's still short, and now it takes advantage of Jax-optimized features like 2D convolution.

Created a really dumb neural network and endowed it with the ability to play itself while following the rules of the game.

The game can be watched on the terminal

All that's left is training the neural network! Will be ~~copying off of~~ taking inspiration from the mnist Python file